chore/dotnet 10 (#540)
- **Migrate to .NET 10**
- **move to slnx**
- **update tools and scripts**

+++ src/Elastic.CommonSchema.BenchmarkDotNetExporter/Domain/BenchmarkLaunchInformation.cs
@@ -32,7 +32,7 @@ namespace Elastic.CommonSchema.BenchmarkDotNetExporter.Domain
 
 		/// <summary></summary>
 		[JsonPropertyName("invocation_count"), DataMember(Name = "invocation_count")]
-		public int InvocationCount { get; set; }
+		public long InvocationCount { get; set; }
 
 		/// <summary></summary>
 		[JsonPropertyName("max_iteration_count"), DataMember(Name = "max_iteration_count")]
